Measuring memory usage

+++ Assets/Src/Utils/StatsResults.cs
@@ -6,10 +6,10 @@ namespace statsresults
     {
         public float timeSpent { get; set; }
         public int nNodesVisited { get; set; }
-        public float maxMemoryUsed { get; set; } 
+        public long maxMemoryUsed { get; set; } 
         public int nMovesUsed { get; set; }
         
-        public StatsResults(float timeSpent, int nNodesVisited, float maxMemoryUsed, int nMovesUsed)
+        public StatsResults(float timeSpent, int nNodesVisited, long maxMemoryUsed, int nMovesUsed)
         {
             this.timeSpent = timeSpent;
             this.nNodesVisited = nNodesVisited;
